Rename and tidy up.
