+++ eth-reference-tests/src/referenceTest/java/tech/pegasys/teku/reference/phase0/bls/BlsVerifyTestExecutor.java
@@ -13,13 +13,11 @@
 
 package tech.pegasys.teku.reference.phase0.bls;
 
-import static java.util.stream.Collectors.toList;
 import static org.assertj.core.api.Assertions.assertThat;
 import static tech.pegasys.teku.ethtests.finder.BlsTestFinder.BLS_DATA_FILE;
 import static tech.pegasys.teku.reference.phase0.TestDataUtils.loadYaml;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
-import java.util.List;
 import org.apache.tuweni.bytes.Bytes;
 import tech.pegasys.teku.bls.BLS;
 import tech.pegasys.teku.bls.BLSPublicKey;
@@ -27,16 +25,17 @@ import tech.pegasys.teku.bls.BLSSignature;
 import tech.pegasys.teku.ethtests.finder.TestDefinition;
 import tech.pegasys.teku.reference.phase0.TestExecutor;
 
-public class BlsFastAggregateVerifyTestType implements TestExecutor {
+public class BlsVerifyTestExecutor implements TestExecutor {
 
   @Override
   public void runTest(final TestDefinition testDefinition) throws Throwable {
     final Data data = loadYaml(testDefinition, BLS_DATA_FILE, Data.class);
-    final List<BLSPublicKey> publicKeys = data.input.getPublicKeys();
+    final BLSPublicKey publicKey = data.input.getPublicKey();
     final Bytes message = data.input.getMessage();
     final BLSSignature signature = data.input.getSignature();
     final boolean expectedResult = data.getOutput();
-    assertThat(BLS.fastAggregateVerify(publicKeys, message, signature)).isEqualTo(expectedResult);
+
+    assertThat(BLS.verify(publicKey, message, signature)).isEqualTo(expectedResult);
   }
 
   private static class Data {
@@ -46,14 +45,18 @@ public class BlsFastAggregateVerifyTestType implements TestExecutor {
     @JsonProperty(value = "output", required = true)
     private boolean output;
 
+    public BLSSignature getSignature() {
+      return input.getSignature();
+    }
+
     public boolean getOutput() {
       return output;
     }
   }
 
   private static class Input {
-    @JsonProperty(value = "pubkeys", required = true)
-    private List<String> publicKeys;
+    @JsonProperty(value = "pubkey", required = true)
+    private String publicKey;
 
     @JsonProperty(value = "message", required = true)
     private String message;
@@ -61,11 +64,8 @@ public class BlsFastAggregateVerifyTestType implements TestExecutor {
     @JsonProperty(value = "signature", required = true)
     private String signature;
 
-    public List<BLSPublicKey> getPublicKeys() {
-      return publicKeys.stream()
-          .map(Bytes::fromHexString)
-          .map(BLSPublicKey::fromBytes)
-          .collect(toList());
+    public BLSPublicKey getPublicKey() {
+      return BLSPublicKey.fromBytes(Bytes.fromHexString(publicKey));
     }
 
     public Bytes getMessage() {
